requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
cachetools>=5.0.0
pytest>=7.0.0
responses>=0.22.0
streamlit>=1.29.0
pandas>=1.5.0
matplotlib>=3.5.0
seaborn>=0.12.0
altair>=4.2.0
pytest-cov>=4.1.0 